backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Nano-Inference POI Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
    "better-profanity>=0.7.0",
    "sentence-transformers[onnx]>=3.0.0",
    "python-dotenv>=1.0.1",
    "diffusers>=0.36.0",
    "transformers>=4.56.2",
    "accelerate>=1.12.0",
    "torch>=2.8.0",
]
